+++ spec/rspec/core/rake_task_spec.rb
@@ -148,6 +148,24 @@ module RSpec::Core
       end
     end
 
+    context "with_clean_environment is set" do
+      it "removes the environment variables", :if => RUBY_VERSION >= '1.9.0', :unless => RSpec::Support::Ruby.jruby? do
+        with_env_vars 'MY_ENV' => 'ABC' do
+          if RSpec::Support::OS.windows?
+            essential_shell_variables = /\["ANSICON", "ANSICON_DEF", "HOME", "TMPDIR", "USER"\]/
+          else
+            essential_shell_variables = /\["PWD"(?:, "SHLVL")?(?:, "_")?(?:, "__CF_USER_TEXT_ENCODING")?\]/
+          end
+
+          expect {
+            task.with_clean_environment = true
+            task.ruby_opts = '-e "puts \"Environment: #{ENV.keys}\""'
+            task.run_task false
+          }.to avoid_outputting.to_stderr.and output(essential_shell_variables).to_stdout_from_any_process
+        end
+      end
+    end
+
     def loaded_files
       args = Shellwords.split(spec_command)
       args -= [task.class::RUBY, "-S", task.rspec_path]